Subsets Sum : Dynamic Programming
Dark Peace is volunteer in VISION 2K17.ChiefCO gave him a task to complete it before the CODEGEEKS which is on 11th march.ChiefCo asks DarkPeace to do the work as soon as possible.So to complete the task as soon as possible Dark Peace asks you to help him.You will be given two set of length N and M.Your task is to find whether the subset is present in the first set whose sum of elements is equal to the member of the set.You have to check for every member of the second set. Print Yes if subset is present and if not present print No and if the member exceeds the maximum sum of the set1 then print -1.